@@ -5,43 +5,44 @@
     "license": "MIT",
     "homepage": "https://teddriggs.github.io/ginscorer",
     "dependencies": {
-        "@types/lodash-es": "4.17.5",
-        "@types/node": "^12.20.37",
-        "@types/react": "^17.0.37",
-        "@types/react-dom": "^17.0.11",
-        "@types/react-redux": "7.1.20",
-        "@types/react-responsive": "8.0.4",
-        "@types/react-transition-group": "4.4.4",
         "classnames": "2.3.1",
         "controlled-input": "0.9.0",
-        "gh-pages": "3.2.3",
         "lodash-es": "4.17.21",
         "memoize-one": "6.0.0",
-        "node-sass": "6.0.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-redux": "7.2.5",
+        "react": "18.1.0",
+        "react-dom": "18.1.0",
+        "react-redux": "8.0.0",
         "react-responsive": "^8.0.3",
-        "react-scripts": "4.0.3",
         "react-transition-group": "4.4.2",
         "redux": "4.1.1",
         "redux-loop": "6.1.0",
         "reselect": "3.0.1",
-        "typescript": "4.4.3",
         "w3c-keys": "1.0.3",
         "web-vitals": "^1.1.2"
     },
     "scripts": {
         "predeploy": "npm run build",
         "deploy": "gh-pages -d build",
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
+        "start": "vite",
+        "build": "vite build",
         "format": "./node_modules/.bin/prettier -u -w --ignore-path=.gitignore"
     },
     "devDependencies": {
-        "prettier": "^2.4.1"
+        "@types/lodash-es": "4.17.7",
+        "@types/node": "20.14.8",
+        "@types/react": "18.3.12",
+        "@types/react-dom": "18.3.1",
+        "@types/react-redux": "7.1.20",
+        "@types/react-responsive": "8.0.4",
+        "@types/react-transition-group": "4.4.4",
+        "@vitejs/plugin-react": "4.7.0",
+        "gh-pages": "3.2.3",
+        "node-sass": "6.0.1",
+        "prettier": "^2.4.1",
+        "sass-embedded": "1.93.3",
+        "typescript": "4.4.3",
+        "vite": "6.4.1",
+        "vite-plugin-pwa": "1.2.0"
     },
     "browserslist": [
         ">0.2%",
